backend/requirements.txt
djangorestframework
gunicorn
psycopg2-binary
django-environ